# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Operator Companion edge-api",
    "version": "0.5.0"
  },
  "paths": {
...
    "/shift/start": {
      "post": {
        "tags": [
          "shift"
        ],
        "summary": "Shift Start",
        "operationId": "shift_start_shift_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShiftOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/shift/end": {
      "post": {
        "tags": [
          "shift"
        ],
        "summary": "Shift End",
        "operationId": "shift_end_shift_end_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EndRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShiftOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/shift/current": {
      "get": {
        "tags": [
          "shift"
        ],
        "summary": "Shift Current",
        "description": "The machine's shift now (ACTIVE, else today's row, else null) + the last handover note\nleft on this machine by a previous shift.",
        "operationId": "shift_current_shift_current_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "machine",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "default: the token's machine",
              "title": "Machine"
            },
            "description": "default: the token's machine"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShiftOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/readiness": {
      "post": {
        "tags": [
          "readiness"
        ],
        "summary": "Submit",
        "operationId": "submit_readiness_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReadinessIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadinessOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EndRequest": {
        "properties": {
          "machine_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Machine Id"
          },
          "handover_note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Handover Note"
          }
        },
        "type": "object",
        "title": "EndRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReadinessIn": {
        "properties": {
          "machine_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Machine Id"
          },
          "sleep_last_24h_h": {
            "type": "number",
            "maximum": 24.0,
            "minimum": 0.0,
            "title": "Sleep Last 24H H"
          },
          "sleep_last_48h_h": {
            "type": "number",
            "maximum": 48.0,
            "minimum": 0.0,
            "title": "Sleep Last 48H H"
          },
          "feel_score": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Feel Score"
          },
          "rt_mean_ms": {
            "type": "number",
            "maximum": 5000.0,
            "exclusiveMinimum": 0.0,
            "title": "Rt Mean Ms"
          },
          "rt_sd_ms": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Rt Sd Ms"
          },
          "rt_lapses": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Rt Lapses"
          },
          "camera_used": {
            "type": "boolean",
            "title": "Camera Used",
            "default": false
          },
          "long_blinks_20s": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Long Blinks 20S"
          }
        },
        "type": "object",
        "required": [
          "sleep_last_24h_h",
          "sleep_last_48h_h",
          "feel_score",
          "rt_mean_ms",
          "rt_lapses"
        ],
        "title": "ReadinessIn"
      },
      "ReadinessOut": {
        "properties": {
          "schema": {
            "type": "string",
            "title": "Schema",
            "default": "readiness.v1"
          },
          "check_id": {
            "type": "string",
            "title": "Check Id"
          },
          "operator_id": {
            "type": "string",
            "title": "Operator Id"
          },
          "shift_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shift Id"
          },
          "ts": {
            "type": "string",
            "title": "Ts"
          },
          "sleep_last_24h_h": {
            "type": "number",
            "title": "Sleep Last 24H H"
          },
          "sleep_last_48h_h": {
            "type": "number",
            "title": "Sleep Last 48H H"
          },
          "feel_score": {
            "type": "integer",
            "title": "Feel Score"
          },
          "rt_mean_ms": {
            "type": "number",
            "title": "Rt Mean Ms"
          },
          "rt_sd_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rt Sd Ms"
          },
          "rt_lapses": {
            "type": "integer",
            "title": "Rt Lapses"
          },
          "rt_delta_vs_baseline_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rt Delta Vs Baseline Pct"
          },
          "baseline_source": {
            "type": "string",
            "title": "Baseline Source"
          },
          "camera_used": {
            "type": "boolean",
            "title": "Camera Used"
          },
          "long_blinks_20s": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Long Blinks 20S"
          },
          "heat_level": {
            "type": "string",
            "title": "Heat Level"
          },
          "score": {
            "type": "integer",
            "title": "Score"
          },
          "rating": {
            "type": "string",
            "title": "Rating"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reasons"
          },
          "supervisor_override": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supervisor Override"
          }
        },
        "type": "object",
        "required": [
          "check_id",
          "operator_id",
          "shift_id",
          "ts",
          "sleep_last_24h_h",
          "sleep_last_48h_h",
          "feel_score",
          "rt_mean_ms",
          "rt_sd_ms",
          "rt_lapses",
          "rt_delta_vs_baseline_pct",
          "baseline_source",
          "camera_used",
          "long_blinks_20s",
          "heat_level",
          "score",
          "rating",
          "reasons"
        ],
        "title": "ReadinessOut",
        "description": "readiness.v1"
      },
...
      "ShiftOut": {
        "properties": {
          "shift": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shift"
          },
          "previous_handover_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Handover Note"
          },
          "as_of": {
            "type": "string",
            "title": "As Of"
          }
        },
        "type": "object",
        "required": [
          "shift",
          "previous_handover_note",
          "as_of"
        ],
        "title": "ShiftOut"
      },
...
      "StartRequest": {
        "properties": {
          "machine_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Machine Id"
          }
        },
        "type": "object",
        "title": "StartRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}